+++ backend/tests/baserow/contrib/database/rows/test_row_webhook_event_types.py
@@ -1,6 +1,8 @@
 import pytest
 
 from baserow.contrib.database.webhooks.registries import webhook_event_type_registry
+from baserow.contrib.database.fields.handler import FieldHandler
+from baserow.contrib.database.rows.handler import RowHandler
 
 
 @pytest.mark.django_db()
@@ -54,10 +56,47 @@ def test_row_created_event_type(data_fixture):
 def test_row_updated_event_type(data_fixture):
     user = data_fixture.create_user()
     table = data_fixture.create_database_table(user=user)
-    field = data_fixture.create_text_field(table=table, primary=True, name="Test 1")
+    table_2 = data_fixture.create_database_table(database=table.database)
+
+    text_field = data_fixture.create_text_field(
+        table=table, primary=True, name="Test 1"
+    )
+    table_2_primary_field = data_fixture.create_text_field(
+        table=table_2, name="Primary Field", primary=True
+    )
+
+    link_row_field = FieldHandler().create_field(
+        user=user,
+        table=table,
+        type_name="link_row",
+        name="Link",
+        link_row_table=table_2,
+    )
+
+    lookup_model = table_2.get_model()
+    i1 = lookup_model.objects.create(
+        **{f"field_{table_2_primary_field.id}": "Lookup 1"}
+    )
+
+    grid = data_fixture.create_grid_view(table=table)
+    data_fixture.create_grid_view_field_option(grid, link_row_field, hidden=False)
 
     model = table.get_model()
-    row = model.objects.create()
+    row = model.objects.create(**{f"field_{text_field.id}": "Old Test value"})
+    getattr(row, f"field_{link_row_field.id}").add(i1.id)
+
+    before_return = webhook_event_type_registry.get(
+        "row.updated"
+    ).get_test_call_before_return(table, row, model)
+
+    row = RowHandler().update_row(
+        user=user,
+        table=table,
+        row_id=row.id,
+        values={f"field_{text_field.id}": "New Test value"},
+    )
+    row.refresh_from_db()
+
     webhook = data_fixture.create_table_webhook(
         table=table,
         request_method="POST",
@@ -70,7 +109,7 @@ def test_row_updated_event_type(data_fixture):
         model=model,
         table=table,
         row=row,
-        before_return=row,
+        before_return=before_return,
     )
     assert payload == {
         "table_id": table.id,
@@ -80,12 +119,14 @@ def test_row_updated_event_type(data_fixture):
         "values": {
             "id": 1,
             "order": "1.00000000000000000000",
-            f"field_{field.id}": None,
+            f"field_{text_field.id}": "New Test value",
+            f"field_{link_row_field.id}": [{"id": 1, "value": "Lookup 1"}],
         },
         "old_values": {
             "id": 1,
             "order": "1.00000000000000000000",
-            f"field_{field.id}": None,
+            f"field_{text_field.id}": "Old Test value",
+            f"field_{link_row_field.id}": [{"id": 1, "value": "Lookup 1"}],
         },
     }
 
@@ -97,7 +138,7 @@ def test_row_updated_event_type(data_fixture):
         model=model,
         table=table,
         row=row,
-        before_return=row,
+        before_return=before_return,
     )
     assert payload == {
         "table_id": table.id,
@@ -107,12 +148,14 @@ def test_row_updated_event_type(data_fixture):
         "values": {
             "id": 1,
             "order": "1.00000000000000000000",
-            f"Test 1": None,
+            f"{text_field.name}": "New Test value",
+            f"{link_row_field.name}": [{"id": 1, "value": "Lookup 1"}],
         },
         "old_values": {
             "id": 1,
             "order": "1.00000000000000000000",
-            f"Test 1": None,
+            f"{text_field.name}": "Old Test value",
+            f"{link_row_field.name}": [{"id": 1, "value": "Lookup 1"}],
         },
     }