+++ changelog.md
@@ -18,6 +18,7 @@
 * Fixed the unchecked percent aggregation calculation
 * Raise Airtable import task error and fixed a couple of minor import bugs.
 * Add loading bar when syncing templates to make it obvious Baserow is still loading.
+* Fixed bug where old values are missing in the update trigger of the webhook.
 
 ## Released (2022-03-03 1.9.1)